Adds section units and splits them by section type
