+++ internal/pkg/section/sections_test.go
@@ -1,54 +1,8 @@
 package section
 
-/*func TestAssertionCopy(t *testing.T) {
-	assertion := GetMessage().Content[0].(*Assertion)
-	aCopy := assertion.Copy(assertion.Context, assertion.SubjectZone)
-	CheckAssertion(assertion, aCopy, t)
-	if assertion == aCopy {
-		t.Error("Assertion was not copied. Pointer is still the same.")
-	}
-}
+/*
 
-func TestShardCopy(t *testing.T) {
-	shard := GetMessage().Content[1].(*Shard)
-	sCopy := shard.Copy(shard.Context, shard.SubjectZone)
-	CheckShard(shard, sCopy, t)
-	if shard == sCopy {
-		t.Error("Assertion was not copied. Pointer is still the same.")
-	}
-}
 
-func TestAssertionInterval(t *testing.T) {
-	var tests = []struct {
-		input *Assertion
-		want  string
-	}{
-		{&Assertion{SubjectName: "test"}, "test"},
-		{new(Assertion), ""},
-	}
-	for i, test := range tests {
-		if test.input.Begin() != test.want || test.input.End() != test.want {
-			t.Errorf("%d: Assertion Begin and End are not as expected=%s actualBegin=%s actualEnd=%s", i, test.want, test.input.Begin(), test.input.End())
-		}
-	}
-}
-
-func TestShardInterval(t *testing.T) {
-	var tests = []struct {
-		input     *Shard
-		wantBegin string
-		wantEnd   string
-	}{
-		{&Shard{RangeFrom: "a", RangeTo: "z"}, "a", "z"},
-		{new(Shard), "", ""},
-	}
-	for i, test := range tests {
-		if test.input.Begin() != test.wantBegin || test.input.End() != test.wantEnd {
-			t.Errorf("%d: Assertion Begin and End are not as expectedBegin=%s expectedEnd=%s actualBegin=%s actualEnd=%s", i,
-				test.wantBegin, test.wantEnd, test.input.Begin(), test.input.End())
-		}
-	}
-}
 
 func TestZoneInterval(t *testing.T) {
 	var tests = []struct {
@@ -64,23 +18,7 @@ func TestZoneInterval(t *testing.T) {
 	}
 }
 
-func TestAssertionHash(t *testing.T) {
-	var tests = []struct {
-		input *Assertion
-		want  string
-	}{
-		{nil, "A_nil"},
-		{new(Assertion), "A____[]_[]"},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx", Content: GetAllValidObjects()[:3],
-			Signatures: []Signature{Signature{PublicKeyID: PublicKeyID{KeySpace: RainsKeySpace, Algorithm: Ed25519}, ValidSince: 1000, ValidUntil: 2000, Data: []byte("SigData")}}},
-			"A_name_zone_ctx_[OT:1 OV:{example.com [3 2]} OT:2 OV:2001:db8:: OT:3 OV:192.0.2.0]_[{KS=0 AT=1 VS=1000 VU=2000 KP=0 data=53696744617461}]"},
-	}
-	for i, test := range tests {
-		if test.input.Hash() != test.want {
-			t.Errorf("%d: Wrong assertion hash. expected=%v, actual=%v", i, test.want, test.input.Hash())
-		}
-	}
-}
+
 
 func TestShardHash(t *testing.T) {
 	var tests = []struct {
@@ -257,60 +195,7 @@ func TestAddressZoneHash(t *testing.T) {
 	}
 }
 
-func TestEqualContextZoneName(t *testing.T) {
-	var tests = []struct {
-		input *Assertion
-		param *Assertion
-		want  bool
-	}{
-		{new(Assertion), nil, false},
-		{new(Assertion), new(Assertion), true},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"}, new(Assertion), false},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"},
-			&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"}, true},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"},
-			&Assertion{SubjectName: "diffname", SubjectZone: "zone", Context: "ctx"}, false},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"},
-			&Assertion{SubjectName: "name", SubjectZone: "diffzone", Context: "ctx"}, false},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"},
-			&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "diffctx"}, false},
-		{&Assertion{SubjectName: "name", SubjectZone: "zone", Context: "ctx"},
-			&Assertion{SubjectName: "diffname", SubjectZone: "diffzone", Context: "diffctx"}, false},
-	}
-	for i, test := range tests {
-		if test.input.EqualContextZoneName(test.param) != test.want {
-			t.Errorf("%d: EqualContextZoneName() returns incorrect result. expected=%v, actual=%v", i, test.want, test.input.EqualContextZoneName(test.param))
-		}
-	}
-}
 
-func TestAssertionString(t *testing.T) {
-	var tests = []struct {
-		input *Assertion
-		want  string
-	}{
-		{nil, "Assertion:nil"},
-		{new(Assertion), "Assertion:[SN= SZ= CTX= CONTENT=[] SIG=[]]"},
-		{
-			&Assertion{
-				SubjectName: "name",
-				SubjectZone: "zone",
-				Context:     "ctx",
-				Content:     GetAllValidObjects()[:3],
-				Signatures: []Signature{
-					Signature{PublicKeyID: PublicKeyID{KeySpace: RainsKeySpace, Algorithm: Ed25519, KeyPhase: 1}, ValidSince: 1000, ValidUntil: 2000, Data: []byte("SigData")},
-					Signature{PublicKeyID: PublicKeyID{KeySpace: RainsKeySpace, Algorithm: Ed25519, KeyPhase: 1}, ValidSince: 3000, ValidUntil: 4000, Data: []byte("SigData2")},
-				},
-			},
-			"Assertion:[SN=name SZ=zone CTX=ctx CONTENT=[OT:1 OV:{example.com [3 2]} OT:2 OV:2001:db8:: OT:3 OV:192.0.2.0] SIG=[{KS=0 AT=1 VS=1000 VU=2000 KP=1 data=53696744617461} {KS=0 AT=1 VS=3000 VU=4000 KP=1 data=5369674461746132}]]",
-		},
-	}
-	for i, test := range tests {
-		if test.input.String() != test.want {
-			t.Errorf("%d: Wrong assertion String(). expected=%v, actual=%v", i, test.want, test.input.String())
-		}
-	}
-}
 
 func TestShardString(t *testing.T) {
 	var tests = []struct {
@@ -580,28 +465,7 @@ func TestNotificationCompareTo(t *testing.T) {
 	}
 }
 
-func TestAssertionCompareTo(t *testing.T) {
-	assertions := sortedAssertions(10)
-	var shuffled []Section
-	for _, a := range assertions {
-		shuffled = append(shuffled, a)
-	}
-	shuffleSections(shuffled)
-	sort.Slice(shuffled, func(i, j int) bool {
-		return shuffled[i].(*Assertion).CompareTo(shuffled[j].(*Assertion)) < 0
-	})
-	for i, a := range assertions {
-		CheckAssertion(a, shuffled[i].(*Assertion), t)
-	}
-	a1 := &Assertion{}
-	a2 := &Assertion{Content: []Object{Object{}}}
-	if a1.CompareTo(a2) != -1 {
-		t.Error("Different content length are not sorted correctly")
-	}
-	if a2.CompareTo(a1) != 1 {
-		t.Error("Different content length are not sorted correctly")
-	}
-}
+
 
 func TestShardCompareTo(t *testing.T) {
 	shards := sortedShards(5)
@@ -731,31 +595,9 @@ func TestAddressQueryCompareTo(t *testing.T) {
 	}
 }
 
-func shuffleSections(sections []Section) {
-	for i := len(sections) - 1; i > 0; i-- {
-		j := rand.Intn(i)
-		sections[i], sections[j] = sections[j], sections[i]
-	}
-}
 
-func TestAssertionSort(t *testing.T) {
-	var tests = []struct {
-		input  []Object
-		sorted []Object
-	}{
-		{
-			[]Object{Object{Type: OTIP4Addr, Value: "192.0.2.0"}, Object{Type: OTName, Value: NameObject{Name: "name", Types: []ObjectType{OTDelegation, OTName}}}},
-			[]Object{Object{Type: OTName, Value: NameObject{Name: "name", Types: []ObjectType{OTName, OTDelegation}}}, Object{Type: OTIP4Addr, Value: "192.0.2.0"}},
-		},
-	}
-	for i, test := range tests {
-		a := &Assertion{Content: test.input}
-		a.Sort()
-		if !reflect.DeepEqual(a.Content, test.sorted) {
-			t.Errorf("%d: Assertion.Sort() does not sort correctly expected=%v actual=%v", i, test.sorted, a.Content)
-		}
-	}
-}
+
+
 
 func TestShardSort(t *testing.T) {
 	var tests = []struct {
